Consolidate common script functions
